move tree to package util
