+++ util/format.go
@@ -6,8 +6,6 @@ import (
 	"strings"
 	"time"
 	"unicode/utf8"
-
-	"github.com/mlange-42/track/tree"
 )
 
 const (
@@ -85,8 +83,8 @@ func Format(str string, repl map[string]string) string {
 }
 
 // TreeFormatter formats trees
-type TreeFormatter[T tree.Named] struct {
-	NameFunc     func(t *tree.MapNode[T], indent int) string
+type TreeFormatter[T Named] struct {
+	NameFunc     func(t *MapNode[T], indent int) string
 	Indent       int
 	prefixNone   string
 	prefixEmpty  string
@@ -95,8 +93,8 @@ type TreeFormatter[T tree.Named] struct {
 }
 
 // NewTreeFormatter creates a new TreeFormatter
-func NewTreeFormatter[T tree.Named](
-	nameFunc func(t *tree.MapNode[T], indent int) string,
+func NewTreeFormatter[T Named](
+	nameFunc func(t *MapNode[T], indent int) string,
 	indent int,
 ) TreeFormatter[T] {
 	return TreeFormatter[T]{
@@ -110,13 +108,13 @@ func NewTreeFormatter[T tree.Named](
 }
 
 // FormatTree formats a tree
-func (f *TreeFormatter[T]) FormatTree(t *tree.MapTree[T]) string {
+func (f *TreeFormatter[T]) FormatTree(t *MapTree[T]) string {
 	sb := strings.Builder{}
 	f.formatTree(&sb, t.Root, 0, false, "")
 	return sb.String()
 }
 
-func (f *TreeFormatter[T]) formatTree(sb *strings.Builder, t *tree.MapNode[T], depth int, last bool, prefix string) {
+func (f *TreeFormatter[T]) formatTree(sb *strings.Builder, t *MapNode[T], depth int, last bool, prefix string) {
 	pref := prefix
 	if depth > 0 {
 		pref = prefix + f.createPrefix(last)
