+++ util/tree.go
@@ -1,4 +1,4 @@
-package tree
+package util
 
 import "fmt"
